==> notebooks/3_final_prep.ipynb <==
# Final preparation

import platform; print(platform.platform())
import sys; print("Python", sys.version)

class cfg:
    seed = 42
    nfolds = 5
    njobs = 2

import numpy as np
import pandas as pd
import os

import matplotlib.pyplot as plt 
import seaborn as sns
sns.set_theme(style='darkgrid', palette='Set1')

from sklearn.ensemble import IsolationForest, HistGradientBoostingRegressor
from sklearn.preprocessing import OneHotEncoder
from sklearn.model_selection import  KFold, train_test_split
from sklearn.metrics import mean_squared_error, mean_absolute_error

## Read data

# load data
train0 = pd.read_csv('../data/raw/train.csv', index_col=0).reset_index(drop=True)
test = pd.read_csv('../data/raw/test.csv', index_col=0).reset_index(drop=True)
train_orig = pd.read_csv('../data/extra/train_orig.csv')

train0.head()

# sample a validation set then add original data to train set
train0['IsSynthetic'] = 1.0
test['IsSynthetic'] = 1.0
train_orig['IsSynthetic'] = 0.0

train = train0.sample(frac = 0.6, random_state=cfg.seed)
valid = train0.drop(train.index)

train = pd.concat([train, train_orig], axis=0)

train.reset_index(drop=True, inplace=True)
valid.reset_index(drop=True, inplace=True)

# quick info
TARGET = 'MedHouseVal'
FEATURES = [c for c in train.columns if c not in [TARGET]]

print(f'Target: {TARGET}\nFeatures: {FEATURES}')
print('Train set shape:', train.shape)
print('Validation set shape:', valid.shape)
print('Test set shape:', test.shape)

## Outliers

# isolation forest for outlier detection
iso = IsolationForest(
    n_estimators=1000,
    max_samples=0.3,
    contamination='auto',
    verbose=0,
    n_jobs=cfg.njobs,
    random_state=cfg.seed
)

# fit isolation forest on synthetic data
_ = iso.fit(train.loc[train['IsSynthetic']==1, FEATURES])

# predict anomaly scores
train['AnomalyScore'] = iso.score_samples(train[FEATURES])
valid['AnomalyScore'] = iso.score_samples(valid[FEATURES])
test['AnomalyScore'] = iso.score_samples(test)

## Cartesian coordinates

# taken from (https://www.kaggle.com/code/dmitryuarov/ps-s3e1-coordinates-key-to-victory)

def crt_crds(df): 
    '''Obtain rotation of the caartesian cordinates'''

    df['rot_15_x'] = (np.cos(np.radians(15)) * df['Longitude']) + (np.sin(np.radians(15)) * df['Latitude'])
    df['rot_15_y'] = (np.cos(np.radians(15)) * df['Latitude']) + (np.sin(np.radians(15)) * df['Longitude'])
    df['rot_30_x'] = (np.cos(np.radians(30)) * df['Longitude']) + (np.sin(np.radians(30)) * df['Latitude'])
    df['rot_30_y'] = (np.cos(np.radians(30)) * df['Latitude']) + (np.sin(np.radians(30)) * df['Longitude'])
    df['rot_45_x'] = (np.cos(np.radians(45)) * df['Longitude']) + (np.sin(np.radians(45)) * df['Latitude'])

    return df

train = crt_crds(train)
valid = crt_crds(valid)
test = crt_crds(test)

from sklearn.decomposition import PCA

def pca_crds(df):
    '''obtain pca coordinates'''
    
    coordinates = df[['Latitude', 'Longitude']]#.values
    pca_obj = PCA().fit(coordinates)
    df['pca_lon'] = pca_obj.transform(df[['Latitude', 'Longitude']])[:,0]
    df['pca_lat'] = pca_obj.transform(df[['Latitude', 'Longitude']])[:,1]
    return df

train = pca_crds(train)
valid = pca_crds(valid)
test = pca_crds(test)

train

# get place from coordinates
import reverse_geocoder as rg

places = [
    'Los Angeles County', 'Orange County', 'Kern County',
    'Alameda County', 'San Francisco County', 'Ventura County',
    'Santa Clara County', 'Fresno County', 'Santa Barbara County',
    'Contra Costa County', 'Yolo County', 'Monterey County',
    'Riverside County', 'Napa County'
]

def place_crds(df):
    '''obtain place from coordinates'''

    coordinates = list(zip(df['Latitude'], df['Longitude']))
    results = rg.search(coordinates)
    df['place'] = [x['admin2'] for x in results]

    def replace(x):
        if x in places:
            return x
        else:
            return 'Other'
    
    df['place'] = df['place'].apply(lambda x: replace(x))

    ohe = OneHotEncoder(handle_unknown='ignore', sparse_output=False)
    places_ohe = ohe.fit_transform(df['place'].values.reshape(-1, 1))
    names = [n.split('_')[1] for n in ohe.get_feature_names_out()]
    tmp = pd.DataFrame(places_ohe, columns=names)

    df = pd.concat([df.drop(columns='place'), tmp], axis=1)

    return df

train = place_crds(train)
valid = place_crds(valid)
test = place_crds(test)

## Save data

# quick info
TARGET = 'MedHouseVal'
FEATURES = [c for c in train.columns if c not in [TARGET]]

print(f'Target: {TARGET}\nFeatures: {FEATURES}')
print('Train set shape:', train.shape)
print('Valid set shape:', valid.shape)
print('Test set shape:', test.shape)

out_path = '../data/final/'
os.makedirs(out_path, exist_ok=True)

train.to_csv(out_path + 'train.csv', index=False)
valid.to_csv(out_path + 'valid.csv', index=False)
test.to_csv(out_path + 'test.csv', index=False)
