==> notebooks/4_lazypredict.ipynb <==
import platform; print(platform.platform())
import sys; print("Python", sys.version)

class cfg:
    seed = 42
    nfolds = 5
    njobs = 2

import numpy as np
import pandas as pd
import os

import lazypredict
from lazypredict.Supervised import LazyRegressor

from sklearn.model_selection import train_test_split

# load data
train = pd.read_csv('../data/final/train.csv')
valid = pd.read_csv('../data/final/valid.csv')
test = pd.read_csv('../data/final/test.csv')

train.head()

# quick info
TARGET = 'MedHouseVal'
FEATURES = [c for c in train.columns if c not in [TARGET]]

print(f'Target: {TARGET}\nFeatures: {FEATURES}')
print('Train set shape:', train.shape)
print('Valid set shape:', valid.shape)
print('Test set shape:', test.shape)

# sample train set
smp = train.sample(10000, random_state=cfg.seed)

# get features and target
x, y = smp[FEATURES], smp[TARGET]
x_val, y_val = valid[FEATURES], valid[TARGET]

# select regressors
list(enumerate(n[0] for n in lazypredict.Supervised.REGRESSORS))
idx = [0, 1, 3, 4, 6, 7, 8, 12, 13, 14, 19, 23, 24 , 25, 26, 33, 35, 36, 37, 40, 41]
mask = [True if est in idx else False for est in range(len(lazypredict.Supervised.REGRESSORS))]

from itertools import compress
lazypredict.Supervised.REGRESSORS = list(compress(lazypredict.Supervised.REGRESSORS, mask))
lazypredict.Supervised.REGRESSORS

# lazypredict
reg = LazyRegressor(verbose=0, ignore_warnings=False, custom_metric=None)
models, predictions = reg.fit(x, x_val, y, y_val)

models
